feat: refactor opportunity components to unify design and remove deprecated elements
