+++ src/web/src/components/Opportunity/OpportunitiesGrid.tsx
@@ -1,8 +1,7 @@
-import { OpportunityPublicSmallComponent } from "./OpportunityPublicSmall";
-import { OpportunityPublicSmallComponentV2 } from "./OpportunityPublicSmallV2";
 import type { OpportunitySearchResultsInfo } from "~/api/models/opportunity";
 import Link from "next/link";
 import { OPPORTUNITY_SEARCH_DESIGN_V2 } from "~/lib/constants";
+import { OpportunityPublicSmallComponent } from "./OpportunityPublicSmall";
 
 interface InputProps {
   [id: string]: any;
@@ -41,7 +40,7 @@ export const OpportunitiesGrid: React.FC<InputProps> = ({
             // NEW: fixed-size cards that wrap to fit as many per row as the screen allows
             <div className="flex w-full flex-row flex-wrap justify-center gap-4">
               {data.items.map((item: any) => (
-                <OpportunityPublicSmallComponentV2
+                <OpportunityPublicSmallComponent
                   key={`${id}_${item.id}`}
                   data={item}
                 />
