+++ src/web/src/pages/opportunities/[[...query]]/index.tsx
@@ -47,8 +47,6 @@ import AnimatedText from "~/components/Opportunity/AnimatedText";
 import FilterTab from "~/components/Opportunity/FilterTab";
 import { OpportunitiesGrid } from "~/components/Opportunity/OpportunitiesGrid";
 import { OpportunityFilterVertical } from "~/components/Opportunity/OpportunityFilterVertical";
-import { OpportunityPublicSmallComponent } from "~/components/Opportunity/OpportunityPublicSmall";
-import { OpportunityPublicSmallComponentV2 } from "~/components/Opportunity/OpportunityPublicSmallV2";
 import {
   getEngagementConfig,
   getTypeConfig,
@@ -73,6 +71,7 @@ import { OpportunityCategoryHorizontalCard } from "~/components/Opportunity/Oppo
 import { CustomSlider } from "~/components/Carousel";
 import Link from "next/link";
 import { IoMdHeart } from "react-icons/io";
+import { OpportunityPublicSmallComponent } from "~/components/Opportunity/OpportunityPublicSmall";
 
 // 👇 SSG
 // This page is statically generated at build time on server-side
@@ -2095,7 +2094,7 @@ const Opportunities: NextPageWithLayout<{
                           loadData={loadDataJobs}
                           totalAll={opportunities_jobs_render.totalCount!}
                           renderSlide={(item, index) => (
-                            <OpportunityPublicSmallComponentV2
+                            <OpportunityPublicSmallComponent
                               key={`opportunities_jobs_${item.id}_${index}`}
                               data={item}
                               variant="job"
@@ -2134,7 +2133,7 @@ const Opportunities: NextPageWithLayout<{
                               loadData={loadDataOpportunitiesForUserCountry}
                               totalAll={opportunities_user_country!.totalCount!}
                               renderSlide={(item, index) => (
-                                <OpportunityPublicSmallComponentV2
+                                <OpportunityPublicSmallComponent
                                   key={`opportunities_user_country_${item.id}_${index}`}
                                   data={item}
                                 />
@@ -2161,7 +2160,7 @@ const Opportunities: NextPageWithLayout<{
                             loadData={loadDataFeatured}
                             totalAll={opportunities_featured_render.totalCount!}
                             renderSlide={(item, index) => (
-                              <OpportunityPublicSmallComponentV2
+                              <OpportunityPublicSmallComponent
                                 key={`opportunities_featured_${item.id}_${index}`}
                                 data={item}
                               />
@@ -2191,7 +2190,7 @@ const Opportunities: NextPageWithLayout<{
                               opportunities_allOpportunities_render.totalCount!
                             }
                             renderSlide={(item, index) => (
-                              <OpportunityPublicSmallComponentV2
+                              <OpportunityPublicSmallComponent
                                 key={`opportunities_newOpportunities_${item.id}_${index}`}
                                 data={item}
                               />
@@ -2221,7 +2220,7 @@ const Opportunities: NextPageWithLayout<{
                               opportunities_mostCompleted_render.totalCount!
                             }
                             renderSlide={(item, index) => (
-                              <OpportunityPublicSmallComponentV2
+                              <OpportunityPublicSmallComponent
                                 key={`opportunities_mostCompleted_${item.id}_${index}`}
                                 data={item}
                               />
@@ -2248,7 +2247,7 @@ const Opportunities: NextPageWithLayout<{
                             loadData={loadDataLearning}
                             totalAll={opportunities_learning_render.totalCount!}
                             renderSlide={(item, index) => (
-                              <OpportunityPublicSmallComponentV2
+                              <OpportunityPublicSmallComponent
                                 key={`opportunities_learning_${item.id}_${index}`}
                                 data={item}
                               />
@@ -2275,7 +2274,7 @@ const Opportunities: NextPageWithLayout<{
                             loadData={loadDataTasks}
                             totalAll={opportunities_tasks_render.totalCount!}
                             renderSlide={(item, index) => (
-                              <OpportunityPublicSmallComponentV2
+                              <OpportunityPublicSmallComponent
                                 key={`opportunities_tasks_${item.id}_${index}`}
                                 data={item}
                               />
@@ -2302,7 +2301,7 @@ const Opportunities: NextPageWithLayout<{
                             loadData={loadDataEvents}
                             totalAll={opportunities_events_render.totalCount!}
                             renderSlide={(item, index) => (
-                              <OpportunityPublicSmallComponentV2
+                              <OpportunityPublicSmallComponent
                                 key={`opportunities_events_${item.id}_${index}`}
                                 data={item}
                               />
